backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.1
openai>=1.0.0
chromadb>=0.4.22
PyPDF2>=3.0.0
python-docx>=1.0.1
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
ruff>=0.1.0 